refactor: address Codacy linting issues
Renames short-named variables to meet minimum identifier length rules,
and splits WallpaperManager.swift into WallpaperManager+Pool.swift and
WallpaperManager+Wallpaper.swift to stay within the 400-line limit.
